% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{print.react_extract_result}
\alias{print.react_extract_result}
\title{Print an extraction result}
\usage{
\method{print}{react_extract_result}(x, ...)
}
\arguments{
\item{x}{A result returned by \code{\link[=react_extract]{react_extract()}}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an extraction result
}
